feat: 弹幕设置

+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/bangumi/info/remote/Section.kt
@@ -1,7 +1,7 @@
 package cn.spacexc.bilibilisdk.sdk.bangumi.info.remote
 
 data class Section(
-    val attr: Int,
+    val attr: Long,
     val episode_id: Long,
     val episode_ids: List<Any>,
     val episodes: List<Episode>,
